requirements.txt
accelerate==1.13.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
authlib==1.6.6
beartype==0.22.9
beautifulsoup4==4.14.2
bitsandbytes==0.49.2
cachetools==6.2.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
cycler==0.12.1
cyclopts==4.4.5
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.0
fakeredis==2.33.0
fastjsonschema==2.21.2
fastmcp==2.14.3
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
gitdb==4.0.12
gitpython==3.1.45
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.7.2
idna==3.11
importlib-metadata==8.7.1
jaraco-classes==3.4.0
jaraco-context==6.1.0
jaraco-functools==4.4.0
jeepney==0.9.0
jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter-core==5.9.1
keyring==25.7.0
kiwisolver==1.5.0
lupa==2.6
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mcp==1.25.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
nbformat==5.10.4
networkx==3.6.1
-e file:///app
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.6.1
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==3.0.1
pathable==0.4.4
pathvalidate==3.3.1
pillow==12.1.1
platformdirs==4.5.0
prometheus-client==0.24.1
psutil==7.2.2
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.12.0
pydocket==0.16.6
pygithub==2.8.1
pygments==2.19.2
pyjwt==2.10.1
pynacl==1.6.0
pyparsing==3.3.2
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pyyaml==6.0.3
redis==7.1.0
referencing==0.36.2
regex==2026.2.28
requests==2.32.5
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.28.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
secretstorage==3.5.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sse-starlette==3.1.2
starlette==0.51.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
traitlets==5.14.3
transformers==5.3.0
triton==3.6.0
typer==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.40.0
websockets==16.0
wrapt==1.17.3
zipp==3.23.0
